requirements/prod.txt
Django==4.2.16
easyocr==1.7.2
gunicorn==20.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
pyTelegramBotAPI==4.23.0